Introduce Stockfish NNUE
Fishtest is a key factor of SF success.

Thanks to Fishtest we have not only greately improved ELO but, even more important, we have enabled a kind of joint development and testing that it is the herat of on open source project like SF.

In case of a chess engine this has never been possible before due to missing a strong and strict testing environment that allows many people to contribute in a safe and coordinate way.

Fishtest is a new way of developing chess engines, something that has never exsisted before.

Co-Authored-By: ppigazzini <pasquale.pigazzini@gmail.com>